Not being the right type after loading from file is a runtime error, not a programmer error.

+++ jhm/config.h
@@ -137,8 +137,8 @@ namespace Jhm {
   };
 
   template <>
-  struct TJsonReader<double> {
-    static double Read(const Base::TJson &entry) {
+  struct TJsonReader<uint64_t> {
+    static uint64_t Read(const Base::TJson &entry) {
       ThrowIfWrongKind(Base::TJson::Number, entry);
       return entry.GetNumber();
     }
